pyproject.toml
[project]
name = "xl-extraction-agent"
version = "0.1.0"
description = ""
authors = [
    {name = "Ausang Shukla",email = "thimmaiah@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12,<4.0"
dependencies = [
    "langchain",
    "langgraph",
    "langchain-community",
    "python-dotenv (>=1.1.1,<2.0.0)",
    "langchain-google-genai",
    "langchain-google-vertexai",
    "pandas (>=2.2.2,<3.0.0)",
    "PyPDF2 (>=3.0.1,<4.0.0)",
    "python-docx (>=1.2.0,<2.0.0)",
    "markdown (>=3.6,<4.0)",
    "beautifulsoup4 (>=4.12.3,<5.0.0)",
    "jinja2 (>=3.1.6,<4.0.0)",
    "openpyxl (>=3.1.5,<4.0.0)"
]

[tool.poetry.group.dev.dependencies]
ipykernel = "^6.29.5"

[tool.poetry.scripts]
start = "src.main:main"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
package-mode = false
